Refactor and restructure Reafactor and restructure the methods using module for the app
