+++ modules/capitalize_decorator.rb
