+++ modules/decorator.rb
