+++ modules/list.rb
@@ -38,4 +38,3 @@ module List
     end
   end
 end
-
