+++ modules/nameable.rb
