+++ modules/rental_opp.rb
@@ -1,0 +1,21 @@
+module RentalOpp
+  def create_rental
+    print "Select a book from the following list by number:\n"
+    list_all_books
+    prompt_user if @books.empty?
+    book_index = gets.chomp.to_i - 1
+
+    print "Select a person from the following list by number:\n"
+    list_all_people
+    prompt_user if @people.empty?
+    person_index = gets.chomp.to_i - 1
+
+    print 'Date [yyyy/mm/dd]: '
+    date = gets.chomp
+
+    rental = Rental.new(date, @people[person_index], @books[book_index])
+    @rentals << rental unless @rentals.include?(rental)
+
+    puts "\nRental created successfully 👍 \n\n"
+  end
+end
