+++ modules/trimmer_decorator.rb
