+++ person.rb
@@ -1,4 +1,4 @@
-require_relative './nameable'
+require_relative './modules/nameable'
 
 class Person < Nameable
   attr_accessor :name, :age, :rentals
